requirements.txt
plexapi>=4.16.1
PyYAML>=6.0.2
requests>=2.32.3